requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
airportsdata==20260208
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.12.1
apache-tvm-ffi==0.1.9
astor==0.8.1
attrs==25.4.0
banks==2.4.1
beautifulsoup4==4.14.3
blake3==1.0.8
boto3==1.42.59
botocore==1.42.59
cachetools==7.0.1
cbor2==5.8.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
compressed-tensors==0.12.2
cryptography==46.0.5
cuda-bindings==13.1.1
cuda-pathfinder==1.4.0
cuda-python==13.1.1
cupy-cuda12x==14.0.1
dataclasses-json==0.6.7
datasets==4.6.1
defusedxml==0.7.1
Deprecated==1.2.18
depyf==0.20.0
dill==0.4.0
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.2
email-validator==2.3.0
faiss-cpu==1.13.2
fastapi==0.135.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.14.0
fastar==0.8.0
filelock==3.25.0
filetype==1.2.0
flashinfer-python==0.5.3
frozenlist==1.8.0
fsspec==2026.2.0
gguf==0.18.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-core==2.5.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gptcache==0.1.44
greenlet==3.3.2
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
ijson==3.5.0
importlib_metadata==8.0.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lark==1.2.2
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.15
llama-index-cli==0.5.3
llama-index-core==0.14.15
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.23
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-faiss==0.5.3
llama-index-workflows==2.15.0
llama-parse==0.6.54
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.26.0
mdurl==0.1.2
mistral_common==1.9.1
model-hosting-container-standards==0.1.13
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
ninja==1.13.0
nltk==3.9.3
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.18.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl==4.4.1
nvidia-cutlass-dsl-libs-base==4.4.1
nvidia-ml-py==13.590.48
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==2.24.0
openai-harmony==0.0.8
opencv-python-headless==4.13.0.92
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.13
outlines==0.1.11
packaging==26.0
pandas==2.3.3
partial-json-parser==0.2.1.1.post7
pillow==12.1.1
platformdirs==4.9.2
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==4.25.8
psutil==7.2.2
py-cpuinfo==9.0.0
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycountry==26.2.16
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pypdf==6.7.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.54.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
sentencepiece==0.2.1
sentry-sdk==2.53.0
setproctitle==1.3.7
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.47
sse-starlette==3.3.2
starlette==0.52.1
striprtf==0.0.26
supervisor==4.3.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tinytag==2.2.0
tokenizers==0.22.2
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.67.3
transformers==4.57.6
triton==3.5.0
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
vllm==0.13.0
watchfiles==1.1.1
websockets==16.0
wheel==0.46.3
wrapt==1.17.3
xformers==0.0.29.post2
xgrammar==0.1.27
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
